+++ CHANGELOG.md
@@ -9,6 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix: Remove the main branch if default branch changes.
 - fix: CKV_GIT_6: "Ensure all commits GPG signed"
 - fix: CKV_GIT_5: "Ensure at least two approving reviews for PRs"
+- feat: bypass pull request allowances in branch protection v3
+
+## [1.3.1] - 2023-11-24
+- fix: removed deprecated contexts attribute
+- showcased the following attributes in complete example visibility,has_issues,has_projects,has_wiki,,is_template,allow_merge_commit,delete_branch_on_merge,has_downloads ,auto_init,archive_on_destroy ,vulnerability_alerts,archived
+- added branch protection V3 in complete example
 
 ## [1.3.0] - 2022-11-21
 ### Changes
@@ -47,7 +53,9 @@ Here following resources were added in this release.
 - Github repository resource
 - Initial commit
 
-[Unreleased]: https://github.com/boldlink/terraform-github-repository/1.3.0...HEAD
+[Unreleased]: https://github.com/boldlink/terraform-github-repository/1.3.1...HEAD
+
+[1.3.1]: https://github.com/boldlink/terraform-github-repository/releases/tag/1.3.1
 [1.3.0]: https://github.com/boldlink/terraform-github-repository/releases/tag/1.3.0
 [1.2.0]: https://github.com/boldlink/terraform-github-repository/releases/tag/1.2.0
 [1.1.0]: https://github.com/boldlink/terraform-github-repository/releases/tag/1.1.0